refactor: clean up DRY violations and dead code
- Extract shared cookie adapter into createRouteClient()
- Remove unused imports, duplicate classes, dead category UI
- Fix checkSetupStatus return type to match API response
- Fix credentials.ts doc bug and add SUPABASE_URL to Vercel error
- Merge duplicate hash-utils imports
